requirements/base.txt
Django==5.2.2
django-environ==0.12.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework_simplejwt==5.5.1
drf-commons==1.0.1
psycopg2-binary==2.9.11
python-decouple==3.8
drf-spectacular==0.29.0
pillow==12.1.0
weasyprint==67.0
requests==2.32.5
boto3==1.42.73
celery==5.6.2
redis==6.4.0
django-celery-beat==2.8.1